+++ reactive-streams/src/main/scala/fs2/interop/reactivestreams/package.scala
@@ -49,7 +49,7 @@ package object reactivestreams {
     *
     * The publisher only receives a subscriber when the stream is run.
     */
-  def fromPublisher[F[_]: Async, A](p: Publisher[A], bufferSize: Long): Stream[F, Chunk[A]] =
+  def fromPublisher[F[_]: Async, A](p: Publisher[A], bufferSize: Int): Stream[F, Chunk[A]] =
     Stream
       .eval(StreamSubscriber[F, A](bufferSize))
       .flatMap(s => s.sub.stream(Sync[F].delay(p.subscribe(s))))
@@ -66,7 +66,7 @@ package object reactivestreams {
   implicit final class PublisherOps[A](val publisher: Publisher[A]) extends AnyVal {
 
     /** Creates a lazy stream from an `org.reactivestreams.Publisher` */
-    def toStreamChunk[F[_]: Async](bufferSize: Long): Stream[F, Chunk[A]] =
+    def toStreamChunk[F[_]: Async](bufferSize: Int): Stream[F, Chunk[A]] =
       fromPublisher(publisher, bufferSize)
 
     /** Creates a lazy stream from an `org.reactivestreams.Publisher` */
